Updated front-end and back-end files for #23 and #21.
Corrected request mappings for user and HCP in their respective
controllers to redirect to the appropriate accessLogs page.

Updated column table headings for consistency.

Added redirection functionality to buttons on index files.

Appropriately checks if the user is a patient when created a list of all
entries or a specified date range.